♻️ movie.service 도메인별 분할 (556 → 40+78+90+218+73줄)
- movie.service.ts (40): 컨트롤러 호환 파사드
- movie-read.service.ts (78): 영화 조회 (getMovieDatas, getMovieDetail, recommend)
- movie-score.service.ts (90): 평점 CRUD (upsert/get/getAverage)
- movie-sync.service.ts (218): 외부 API + DB 동기화 (KOFIC/KMDB/TMDB)
  ※ 200줄 룰 9% 초과, sync 오케스트레이션이 한 도메인이라 더 분할 어려움
- movie.formatter.ts (73): convertMovieDataWithCounts, convertKobisMovieData
- 빈 placeholder 파일 movie.datasource.ts, movie.repository.ts 제거
